requirements.txt
pika #A Python library for working with RabbitMQ.
python-dotenv